+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/type/ContinuousProfilingMonitoringProcess.java
@@ -18,21 +18,18 @@
 
 package org.apache.skywalking.oap.server.core.query.type;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.apache.skywalking.oap.server.core.profiling.continuous.storage.ContinuousProfilingTargetType;
 
 import java.util.List;
 
 @Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ContinuousProfilingPolicyTarget {
-    private ContinuousProfilingTargetType type;
-    private List<ContinuousProfilingPolicyItem> checkItems;
+public class ContinuousProfilingMonitoringProcess {
+
+    private String id;
+    private String name;
+    private String detectType;
+    private List<String> labels;
     private int triggeredCount;
     private Long lastTriggerTimestamp;
+
 }