+++ docs/en/changes/changes.md
@@ -46,6 +46,8 @@
 * Filter out unknown_cluster metric data.
 * Support RabbitMQ Monitoring.
 * Support Redis slow logs collection.
+* Fix data loss when query continuous profiling task record.
+* Adapt the continuous profiling task query GraphQL.
 
 #### UI
 * Revert: cpm5d function. This feature is cancelled from backend.